client/requirements.txt
requests==2.25.1
pycrypto==2.6.1